update based on feedback

+++ src/listener/builtin.ts
@@ -24,11 +24,21 @@ export class TokenRevocationListeners {
       isEnterpriseInstall: isEnterpriseInstall,
       teamId: context.teamId,
       enterpriseId: context.enterpriseId,
+      userId: context.userId,
     };
 
-    this.installationStore.deleteInstallation(installQuery, this.logger);
-
-    // add logic to delete bot?
+    if (this.installationStore.deleteInstallation) {
+      this.installationStore.deleteInstallation(installQuery, this.logger);
+      /**
+       * deleteInstallation with enterprise/team_id + user_id -> delete only user tokens
+        deleteBot with enterprise/team_id -> delete only bot tokens
+        deleteAll with enterprise/team_id -> delete both bot and user tokens
+      */
+    } else {
+      throw new Error(
+        `Custom InstallationStore must have deleteInstallation method implemented`,
+      );
+    }
   }
 
   public handleAppUninstalledEvents(context: Context) {
@@ -38,10 +48,20 @@ export class TokenRevocationListeners {
       isEnterpriseInstall: isEnterpriseInstall,
       teamId: context.teamId,
       enterpriseId: context.enterpriseId,
+      userId: context.userId,
     };
-    
-    this.installationStore.deleteInstallation(installQuery, this.logger);
 
-    // add logic to delete bot?
+    if (this.installationStore.deleteInstallation) {
+      this.installationStore.deleteInstallation(installQuery, this.logger);
+      /**
+       * deleteInstallation with enterprise/team_id + user_id -> delete only user tokens
+        deleteBot with enterprise/team_id -> delete only bot tokens
+        deleteAll with enterprise/team_id -> delete both bot and user tokens
+      */
+    } else {
+      throw new Error(
+        `Custom InstallationStore must have deleteInstallation method implemented`,
+      );
+    }
   }
 }
